Rename src/util/Github.ts -> src/util/GithubActions.ts
