+++ src/util/github/PullRequestHelper.test.ts
@@ -1,6 +1,6 @@
 import { setFailed } from "@actions/core"
 import { Context } from "@actions/github/lib/context"
-import { Octokit } from "../Github"
+import { Octokit } from "../GithubActions"
 import PullRequestHelper from "./PullRequestHelper"
 
 jest.mock("@actions/core")
